refactor: extract PermissionChecker class from executor.ts
